flask_api/requirements.txt

flask
flask-sqlalchemy
flask-wtf
flask-talisman
PyJWT
WTForms
psycopg2-binary

# SDSP_Services
# wheel==0.37.1
# bcrypt==3.1.7
# certifi==2022.12.7 #2020.12.5
# cffi==1.14.0
# chardet==3.0.4
# click==7.1.2
# cryptography==2.9
# Flask
# Flask-Bcrypt==0.7.1
# Flask-Login==0.5.0
# Flask-SQLAlchemy==2.4.3

# Flask-WTF==0.14.2
# jyserver==0.0.5
# gunicorn==20.0.4
# gevent==22.10.2
# idna==2.10
# itsdangerous==1.1.0
# Jinja2==2.11.2
# MarkupSafe==1.1.1
# paho-mqtt==1.5.1
# psycopg2==2.8.5
# pycparser==2.20

# requests==2.24.0
# six==1.14.0
# SQLAlchemy==1.3.18
# urllib3==1.25.11
# Werkzeug==0.16.1

# xlrd==1.2.0

